changed name on api.js to index.js. added code to show all the jackets, and did some shanges in styles.css
