+++ index.html
@@ -8,9 +8,7 @@
         <title>Rainy Days homepage</title>
         <meta name="description" content="This is a homepage that sells high quality rain jackets for all type of weather and outdoor activities">
         <script src="https://kit.fontawesome.com/727f4fec37.js" crossorigin="anonymous"></script>
-        <script defer type="module" src="./js/api.js"></script>
-        <script defer type="module" src="js/product.js"></script>
-        <script defer type="module" src="js/spesificproduct.js"></script>
+        <script defer type="module" src="js/index.js"></script>
     </head>
     <body>
         <header>
@@ -61,7 +59,7 @@
             </section>  
                 <h1 class="best-sellers-header">BEST SELLERS</h1>
             <section class="best-sellers">
-                <div class="loader">Loading...</div>
+                <div id="best-sellers-jackets"></div>
             </section>
             <section class="newsletter">
                     <p class="newsletter-text">Sign up for our latest news<br>and exclusive drops & get<br>10% off your first order</p>
