+++ product.html
@@ -8,7 +8,7 @@
         <title>Shop </title>
         <meta name="description" content="Here you can see all the rain jackets that Rainy Days are selling on this web page, and the price of each jacket">
         <script src="https://kit.fontawesome.com/727f4fec37.js" crossorigin="anonymous"></script>
-        <script defer type="module" src="js/api.js"></script>
+        <script defer type="module" src="js/products.js"></script>
     </head>
     <body>
         <header>
